+++ CotEditor/Sources/Settings Window/Panes/ModeSettingsView.swift
@@ -187,6 +187,8 @@ private struct ModeOptionsView: View {
                            isOn: $options.automaticQuoteSubstitution)
                     Toggle(String(localized: "Smart dashes", table: "ModeSettings", comment: "use localization provided by Apple"),
                            isOn: $options.automaticDashSubstitution)
+                    Toggle(String(localized: "Add period with double-space", table: "ModeSettings", comment: "use localization provided by Apple"),
+                           isOn: $options.automaticPeriodSubstitution)
                     Toggle(String(localized: "Automatically insert closing brackets and quotes", table: "ModeSettings"),
                            isOn: $options.automaticSymbolBalancing)
                 }